pyproject.toml
[project]
name = "local-guide-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.12",
    "crawl4ai>=0.6.3",
    "datasets>=3.6.0",
    "faker>=18.13.0",
    "fastapi>=0.115.9",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-openai>=0.3.24",
    "matplotlib>=3.10.3",
    "openai>=1.88.0",
    "pandas>=2.0.0",
    "playwright>=1.52.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
    "sentence-transformers>=4.1.0",
    "sentencepiece>=0.2.0",
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "uvicorn>=0.34.3",
]
